+++ front/src/main/java/ru/yandex/practicum/model/AccountInfo.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -13,5 +15,5 @@ public class AccountInfo {
 
     private String name;
     private String birthdate;
-    private int balance;
+    private BigDecimal balance;
 }
